+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 - resource `storage_zone`: S3 support;
+- resource `compute_container_app`: container `image_digest` attribute;
 ### Fixed
 - resource `compute_container_app`: normalize endpoint protocol casing;